requirements.txt
asgiref==3.4.1
Django==3.2
django-debug-toolbar==3.2.2
django-filter==22.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
uWSGI==2.0.19.1
# channels==3.0.4
# channels-redis==3.3.1
drf-yasg==1.20.0
django-redis==5.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.13.0
